refactor: standardize package name to py_omop2neo4j_lpg (#70)
This change addresses the inconsistency in the package naming.

The package directory `src/omop2neo4j_lpg` has been renamed to `src/py_omop2neo4j_lpg`.

The `pyproject.toml` file has been updated to reflect this change in the `packages` and `scripts` sections.

All import statements in the source and test files have been updated to use the new package name.
