+++ tests/test_queries.py
@@ -1,5 +1,5 @@
 import pytest
-from omop2neo4j_lpg import extraction, loading
+from py_omop2neo4j_lpg import extraction, loading
 
 # --- Tests for extraction.py ---
